Enum egg_mode::error::Error

pub enum Error {
    InvalidResponse,
    MissingValue(&'static str),
    TwitterError(TwitterErrors),
    BadStatus(StatusCode),
    NetError(Error),
    IOError(Error),
    JSONError(ParserError),
    DecodeError(DecoderError),
}

A set of errors that can occur when interacting with Twitter.

Variants
InvalidResponse

The response from Twitter was formatted incorrectly or in an unexpected manner.

MissingValue(&'static str)

The response from Twitter was missing an expected value. The enclosed value was the expected parameter.

TwitterError(TwitterErrors)

The response from Twitter returned an error structure instead of the expected response. The enclosed value was the response from Twitter.

BadStatus(StatusCode)

The response from Twitter gave a response code that indicated an error. The enclosed value was the response code.

NetError(Error)

The web request experienced an error. The enclosed value was returned from hyper.

IOError(Error)

An error was experienced while processing the response stream. The enclosed value was returned from libstd.

JSONError(ParserError)

An error occurred while parsing the JSON resposne. The enclosed value was returned from rustc_serialize.

DecodeError(DecoderError)

An error occurred while loading the JSON response. The enclosed value was returned from rustc_serialize.