Function egg_mode::access_token

pub fn access_token<S: Into<String>>(con_token: &Token, request_token: &Token, verifier: S) -> Result<(Token<'static>, i64, String), Error>

With the given OAuth tokens and verifier, ask Twitter for an access Token that can be used to sign further requests to the Twitter API.

This can be considered Step 3 in obtaining access to a user's account. The Token this function returns represents the user's authorization that your app can use their account, and needs to be given to all other functions in the Twitter API.

The OAuth Verifier this function takes is either given as a result of the OAuth Callback given to request_token, or the PIN given to the user as a result of their access of the authorize_url.

This function also returns the User ID and Username of the authenticated user.