Function egg_mode::authenticate_url

pub fn authenticate_url(request_token: &Token) -> String

With the given request Token, return a URL to redirect a user to so they can accept or reject an authorization request.

This can be considered Step 2 in obtaining access to a user's account. Using the "Sign in with Twitter" authenication flow for websites, your application can redirect the user to the URL returned by this function. Upon accepting the request, the user is redirected to the callback URL given to access_token, with an oauth_token and oauth_verifier appended as a query string. That Verifier can then be given to access_token to complete authorization.