Function egg_mode::authorize_url

pub fn authorize_url(request_token: &Token) -> String

With the given request Token, return a URL that a user can access to accept or reject an authorization request.

This can be considered Step 2 in obtaining access to a user's account. Using PIN-Based Auth for desktop applications, give the URL that this function returns to the user so they can process the authorization request. They will receive a PIN in return, that can be given as the Verifier to access_token.