Function egg_mode::request_token

pub fn request_token<S: Into<String>>(con_token: &Token,
                                      callback: S)
                                      -> Result<Token<'static>, Error>

With the given consumer Token, ask Twitter for a request Token that can be used to request access to the user's account.

This can be considered Step 1 in obtaining access to a user's account. With this Token, a web-based application can use authenticate_url, and a desktop-based application can use authorize_url to perform the authorization request.

The parameter callback is used to provide an OAuth Callback URL for a web- or mobile-based application to receive the results of the authorization request. To use the PIN-Based Auth request, this must be set to "oob". The resulting Token can be passed to authorize_url to give the user a means to accept the request.