Struct egg_mode::Response

pub struct Response<T> {
    pub rate_limit: i32,
    pub rate_limit_remaining: i32,
    pub rate_limit_reset: i32,
    pub response: T,
}

A helper struct to wrap response data with accompanying rate limit information.

Fields
rate_limit: i32

The rate limit ceiling for the given request.

rate_limit_remaining: i32

The number of requests left for the 15-minute window.

rate_limit_reset: i32

The UTC Unix timestamp at which the rate window resets.

response: T

The decoded response from the request.