Module egg_mode::tweet

Structs and functions for working with statuses and timelines.

In this module, you can find various structs and methods to load and interact with tweets and their metadata. This also includes loading a user's timeline, posting a new tweet, or liking or retweeting another tweet. However, this does not include searching for tweets; that functionality is in the search module.

Types

Tweet/TweetEntities/ExtendedTweetEntities: At the bottom of it all, this is the struct that represents a single tweet. The *Entities structs contain information about media, links, and hashtags within their parent tweet.

DraftTweet: This is what you use to post a new tweet. At present, not all available options are supported, but basics like marking the tweet as a reply and attaching a location coordinate are available.

Timeline: Returned by several functions in this module, this is how you cursor through a collection of tweets. See the struct-level documentation for details.

Functions

User actions

These functions perform actions on their given tweets. They require write access to the authenticated user's account.

delete (for creating a tweet, see DraftTweet)

like/unlike

retweet/unretweet

Metadata lookup

These functions either perform some direct lookup of specific tweets, or provide some metadata about the given tweet in a direct (non-Timeline) fashion.

show

lookup/lookup_map (for the differences between these functions, see their respective documentations.)

retweeters_of

retweets_of

Timeline cursors

These functions return Timelines and can be cursored around in the same way. See the documentation for Timeline to learn how to navigate these return values. This correspond to a user's own view of Twitter, or with feeds you might see attached to a user's profile page.

home_timeline/mentions_timeline/retweets_of_me

user_timeline/liked_by

Structs

DraftTweet
Represents an in-progress tweet before it is sent.

ExtendedTweetEntities
Container for extended media information for a tweet.

Timeline
Helper struct to navigate collections of tweets by requesting tweets older or newer than certain IDs.

Tweet
Represents a single status update.

TweetEntities
Container for URL, hashtag, mention, and media information associated with a tweet.

Functions

delete
Delete the given tweet. The authenticated user must be the user who posted the given tweet.

home_timeline
Make a Timeline struct for navigating the collection of tweets posted by the authenticated user and the users they follow.

like
Like the given status as the authenticated user.

liked_by
Make a Timeline struct for navigating the collection of tweets liked by the given user.

lookup
Lookup tweet information for the given list of tweet IDs.

lookup_map
Lookup tweet information for the given list of tweet IDs, and return a map indicating which IDs couldn't be found.

mentions_timeline
Make a Timeline struct for navigating the collection of tweets that mention the authenticated user's screen name.

retweet
Retweet the given status as the authenticated user.

retweeters_of
Lookup the user IDs that have retweeted the given tweet.

retweets_of
Lookup the most recent 100 (or fewer) retweets of the given tweet.

retweets_of_me
Make a Timeline struct for navigating the collection of tweets posted by the authenticated user that have been retweeted by others.

show
Lookup a single tweet by numeric ID.

unlike
Clears a like of the given status as the authenticated user.

unretweet
Unretweet the given status as the authenticated user.

user_timeline
Make a Timeline struct for navigating the collection of tweets posted by the given user, optionally including or excluding replies or retweets.