Function egg_mode::tweet::lookup_map

pub fn lookup_map(
    ids: &[i64], 
    con_token: &Token, 
    access_token: &Token
) -> WebResponse<HashMap<i64, Option<Tweet>>>

Lookup tweet information for the given list of tweet IDs, and return a map indicating which IDs couldn't be found.

This function differs from lookup in how it handles protected or nonexistent tweets. lookup_map returns a map containing every ID in the input slice; tweets that don't exist or can't be read by the authenticated user store None in the map, whereas tweets that could be loaded store Some and the requested status.