Function egg_mode::tweet::mentions_timeline

pub fn mentions_timeline<'a>(con_token: &'a Token,
                             access_token: &'a Token)
                             -> Timeline<'a>

Make a Timeline struct for navigating the collection of tweets that mention the authenticated user's screen name.

This method has a default page size of 20 tweets, with a maximum of 200.

Twitter will only return the most recent 800 tweets by navigating this method.