Struct egg_mode::tweet::TweetEntities

pub struct TweetEntities {
    pub hashtags: Vec<HashtagEntity>,
    pub symbols: Vec<HashtagEntity>,
    pub urls: Vec<UrlEntity>,
    pub user_mentions: Vec<MentionEntity>,
    pub media: Option<Vec<MediaEntity>>,
}

Container for URL, hashtag, mention, and media information associated with a tweet.

If a tweet has no hashtags, financial symbols ("cashtags"), links, or mentions, those respective Vecs will be empty. If there is no media attached to the tweet, that field will be None.

Note that for media attached to a tweet, this struct will only contain the first image of a photo set, or a thumbnail of a video or GIF. Full media information is available in the tweet's extended_entities field.

Fields
hashtags: Vec<HashtagEntity>

Collection of hashtags parsed from the tweet.

symbols: Vec<HashtagEntity>

Collection of financial symbols, or "cashtags", parsed from the tweet.

urls: Vec<UrlEntity>

Collection of URLs parsed from the tweet.

user_mentions: Vec<MentionEntity>

Collection of user mentions parsed from the tweet.

media: Option<Vec<MediaEntity>>

If the tweet contains any attached media, this contains a collection of media information from the tweet.