Enum egg_mode::user::Connection

pub enum Connection {
    None,
    FollowingRequested,
    FollowingReceived,
    FollowedBy,
    Following,
    Blocking,
    Muting,
}

Represents the ways a target account can be connected to another account.

Variants
None

The target account has no relation.

FollowingRequested

The authenticated user has requested to follow the target account.

FollowingReceived

The target account has requested to follow the authenticated user.

FollowedBy

The target account follows the authenticated user.

Following

The authenticated user follows the target account.

Blocking

The authenticated user has blocked the target account.

Muting

The authenticated user has muted the target account.