Function egg_mode::user::followers_ids

pub fn followers_ids<'a, T: Into<UserID<'a>>>(acct: T, con_token: &'a Token, access_token: &'a Token) -> CursorIter<'a, IDCursor>

Lookup the users that follow a given account, but only return their user IDs.

This function returns an iterator over the User IDs returned by Twitter. This method defaults to returning 500 IDs in a single network call; the maximum is 5000.

Choosing only to load the user IDs instead of the full user information results in a call that can return more accounts per-page, which can be useful if you anticipate having to page through several results and don't need all the user information.