Function egg_mode::user::lookup

pub fn lookup<'a, T: 'a>(accts: &'a [T], con_token: &Token, access_token: &Token) -> WebResponse<Vec<TwitterUser>> where &'a T: Into<UserID<'a>>

Lookup a set of Twitter users by either ID and screen name, as applicable.

This function is set up so it can be called with any number of slice types; whether just IDs, just screen names, or even a mix of both (by using &[UserID] directly).

Examples

let mut list: Vec<i64> = Vec::new();

list.push(1234);
list.push(2345);

let users = egg_mode::user::lookup(&list, &con_token, &access_token).unwrap();

let mut list: Vec<String> = Vec::new();

list.push("rustlang".into());
list.push("ThisWeekInRust".into());

let users = egg_mode::user::lookup(&list, &con_token, &access_token).unwrap();

let mut list: Vec<egg_mode::user::UserID> = Vec::new();

list.push(1234.into());
list.push("rustlang".into());

let users = egg_mode::user::lookup(&list, &con_token, &access_token).unwrap();