Function egg_mode::user::mutes

pub fn mutes<'a>(con_token: &'a Token, access_token: &'a Token) -> CursorIter<'a, UserCursor>

Lookup the users that have been muted by the authenticated user.

Note that while loading a user's mutes list is a cursored search, it does not allow you to set the page size. Calling with_page_size on the iterator returned by this function will not change the page size used by the network call. Setting page_size manually may result in an error from Twitter.