Function egg_mode::user::update_follow

pub fn update_follow<'a, T>(acct: T, notifications: Option<bool>, retweets: Option<bool>, con_token: &Token, access_token: &Token) -> WebResponse<Relationship> where T: Into<UserID<'a>>

Update notification settings and reweet visibility for the given user.

Calling this for an account the authenticated user does not already follow will not cause them to follow that user. It will return an error if you pass Some(true) for notifications or Some(false) for retweets. Any other combination of arguments will return a Relationship as if you had called relation between the authenticated user and the given user.