Module egg_mode::cursor

Types and traits to navigate cursored collections.

Much of this module can be considered an implementation detail; the main intended entry point to this code is CursorIter, and that can just be used as an iterator to ignore the rest of this module. The rest of it is available to make sure consumers of the API can understand precisely what types come out of functions that return CursorIter.

Structs

CursorIter
Represents a paginated list of results, such as the users who follow a specific user or the lists owned by that user.

IDCursor
Represents a single-page view into a list of user IDs.

UserCursor
Represents a single-page view into a list of users.

Traits

Cursor
Trait to generalize over paginated views of API results.