Struct egg_mode::search::SearchBuilder

#[must_use = "SearchBuilder is lazy and won\'t do anything unless `call`ed"]

pub struct SearchBuilder<'a> { /* fields omitted */ }

Represents a tweet search query before being sent.

Methods

impl<'a> SearchBuilder<'a>

pub fn lang(self, lang: &'a str) -> Self

Restrict search results to those that have been machine-parsed as the given two-letter language code.

pub fn result_type(self, result_type: ResultType) -> Self

Specify the type of search results to include. The default is Recent.

pub fn count(self, count: u32) -> Self

Set the number of tweets to return per-page, up to a maximum of 100. The default is 15.

pub fn until(self, year: u32, month: u32, day: u32) -> Self

Returns tweets created before the given date. Keep in mind that search is limited to the last 7 days of results, so giving a date here that's older than a week will return no results.

pub fn geocode(self, latitude: f32, longitude: f32, radius: Distance) -> Self

Restricts results to users located within the given radius of the given coordinate. This is preferably populated from location-tagged tweets, but can be filled in from the user's profile as a fallback.

pub fn call(
self,
con_token: &Token,
access_token: &Token
) -> WebResponse<SearchResult<'a>>

Finalize the search terms and return the first page of responses.