Struct egg_mode::search::SearchResult

pub struct SearchResult<'a> {
    pub statuses: Vec<Tweet>,
    pub query: String,
    // some fields omitted
}

Represents a page of search results, along with metadata to request the next or previous page.

Fields
statuses: Vec<Tweet>

The list of statuses in this page of results.

query: String

The query used to generate this page of results. Note that changing this will not affect the next_page method.

Methods

impl<'a> SearchResult<'a>

fn older(&self,
con_token: &Token,
access_token: &Token)
-> WebResponse<SearchResult>

Load the next page of search results for the same query.

fn newer(&self,
con_token: &Token,
access_token: &Token)
-> WebResponse<SearchResult>

Load the previous page of search results for the same query.