Function egg_mode::tweet::user_timeline

pub fn user_timeline<'a, T: Into<UserID<'a>>>(acct: T,
                                              with_replies: bool,
                                              with_rts: bool,
                                              con_token: &'a Token,
                                              access_token: &'a Token)
                                              -> Timeline<'a>

Make a Timeline struct for navigating the collection of tweets posted by the given user, optionally including or excluding replies or retweets.

Attempting to load the timeline of a protected account will only work if the account is the authenticated user's, or if the authenticated user is an approved follower of the account.

This method has a default page size of 20 tweets, with a maximum of 200. Note that asking to leave out replies or retweets will generate pages that may have fewer tweets than your requested page size; Twitter will load the requested number of tweets before removing replies and/or retweets.

Twitter will only load the most recent 3,200 tweets with this method.