Struct egg_mode::tweet::DraftTweet

pub struct DraftTweet<'a> { /* fields omitted */ }

Represents an in-progress tweet before it is sent.

This is your entry point to posting new tweets to Twitter. To begin, make a new DraftTweet by calling new with your desired status text:

use egg_mode::tweet::DraftTweet;

let draft = DraftTweet::new("This is an example status!");

As-is, the draft won't do anything until you call send to post it:

draft.send(&con_token, &access_token).unwrap();

Right now, the options for adding metadata to a post are pretty sparse. See the adaptor functions below to see what metadata can be set. For example, you can use in_reply_to to create a reply-chain like this:

use egg_mode::tweet::DraftTweet;

let draft = DraftTweet::new("I'd like to start a thread here.");
let tweet = draft.send(&con_token, &access_token).unwrap();

let draft = DraftTweet::new("You see, I have a lot of things to say.")
                       .in_reply_to(tweet.response.id);
let tweet = draft.send(&con_token, &access_token).unwrap();

let draft = DraftTweet::new("Thank you for your time.")
                       .in_reply_to(tweet.response.id);
let tweet = draft.send(&con_token, &access_token).unwrap();

Methods

impl<'a> DraftTweet<'a>

pub fn new(text: &'a str) -> Self

Creates a new DraftTweet with the given status text.

pub fn in_reply_to(self, in_reply_to: i64) -> Self

Marks this draft tweet as replying to the given status ID.

Note that this will only properly take effect if the user who posted the given status is @mentioned in the status text, or if the given status was posted by the authenticated user.

pub fn coordinates(self, latitude: f32, longitude: f32, display: bool) -> Self

Attach a lat/lon coordinate to this tweet, and mark whether a pin should be placed on the exact coordinate when the tweet is displayed.

pub fn send(
&self,
con_token: &Token,
access_token: &Token
) -> WebResponse<Tweet>

Send the assembled tweet as the authenticated user.