Struct egg_mode::user::Relationship

pub struct Relationship {
    pub target: RelationTarget,
    pub source: RelationSource,
}

Represents relationship settings between two Twitter accounts.

Fields
target: RelationTarget

Contains settings from the perspective of the target account.

source: RelationSource

Contains settings from the perspective of the source account.

This contains more information than target if the source account is the same as the authenticated user. See the RelationSource page for details.